api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "business-analyst-api"
version = "0.1.0"
description = "AI-powered Business Analyst API"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "google-generativeai==0.3.0",
    "pandas==2.1.3",
    "numpy==1.26.2",
    "matplotlib==3.8.2",
    "seaborn==0.13.0",
    "scipy==1.11.4",
    "python-multipart==0.0.6",
    "aiofiles==23.2.1",
    "corsheaders==4.3.1",
]
